// frontend/src/api/models/PageResponseUsersRoleResponse.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Task Tracker API
 * Task Tracker API
 *
 * The version of the OpenAPI document: 1.0
 * Contact: trnka.roman13@gmail.com
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
import type { UsersRoleResponse } from './UsersRoleResponse';
import {
    UsersRoleResponseFromJSON,
    UsersRoleResponseFromJSONTyped,
    UsersRoleResponseToJSON,
} from './UsersRoleResponse';

/**
 * 
 * @export
 * @interface PageResponseUsersRoleResponse
 */
export interface PageResponseUsersRoleResponse {
    /**
     * 
     * @type {Array<UsersRoleResponse>}
     * @memberof PageResponseUsersRoleResponse
     */
    data?: Array<UsersRoleResponse>;
    /**
     * 
     * @type {number}
     * @memberof PageResponseUsersRoleResponse
     */
    pageNum?: number;
    /**
     * 
     * @type {number}
     * @memberof PageResponseUsersRoleResponse
     */
    pageSize?: number;
    /**
     * 
     * @type {number}
     * @memberof PageResponseUsersRoleResponse
     */
    count?: number;
    /**
     * 
     * @type {number}
     * @memberof PageResponseUsersRoleResponse
     */
    totalPages?: number;
}

/**
 * Check if a given object implements the PageResponseUsersRoleResponse interface.
 */
export function instanceOfPageResponseUsersRoleResponse(value: object): value is PageResponseUsersRoleResponse {
    return true;
}

export function PageResponseUsersRoleResponseFromJSON(json: any): PageResponseUsersRoleResponse {
    return PageResponseUsersRoleResponseFromJSONTyped(json, false);
}

export function PageResponseUsersRoleResponseFromJSONTyped(json: any, ignoreDiscriminator: boolean): PageResponseUsersRoleResponse {
    if (json == null) {
        return json;
    }
    return {
        
        'data': json['data'] == null ? undefined : ((json['data'] as Array<any>).map(UsersRoleResponseFromJSON)),
        'pageNum': json['pageNum'] == null ? undefined : json['pageNum'],
        'pageSize': json['pageSize'] == null ? undefined : json['pageSize'],
        'count': json['count'] == null ? undefined : json['count'],
        'totalPages': json['totalPages'] == null ? undefined : json['totalPages'],
    };
}

export function PageResponseUsersRoleResponseToJSON(value?: PageResponseUsersRoleResponse | null): any {
    if (value == null) {
        return value;
    }
    return {
        
        'data': value['data'] == null ? undefined : ((value['data'] as Array<any>).map(UsersRoleResponseToJSON)),
        'pageNum': value['pageNum'],
        'pageSize': value['pageSize'],
        'count': value['count'],
        'totalPages': value['totalPages'],
    };
}